move contracts to another path
