+++ test/ethereum/deploy/_NAME_/contracts/_NAME_/names.json
@@ -1,0 +1,4 @@
+[
+  "bar",
+  "foo"
+]
