+++ test/ethereum/deploy/_NAME_/contracts/contracts.bats
